mvoed db funcs out
